Enhance HeroSection with language toggle and improve Navbar background gradient
