+++ iiitn-website/src/components/Navbar.tsx
@@ -8,34 +8,18 @@ const imgSrc = logo;
 const Navbar = () => {
 	const [isScrolled, setIsScrolled] = useState(false);
 	const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
-	const [isHomePage,setIsHomePage] = useState(window.location.pathname === "/");
+	const [isHomePage, setIsHomePage] = useState(window.location.pathname === "/");
 	const [openDropdown, setOpenDropdown] = useState<number | null>(null);
-	
+
 	const [openSubmenu, setOpenSubmenu] = useState<number | null>(null);
 	const [openNestedSubmenu, setOpenNestedSubmenu] = useState<number | null>(
 		null
 	);
 
 	useEffect(() => {
-		const toggleLanguage = () => {
-			setTimeout(() => {
-				setIsEnglish(true);
-				setTimeout(() => {
-					setIsEnglish(false);
-					setTimeout(() => {
-						setIsEnglish(true);
-					}, 5000);
-				}, 5000);
-			}, 5000);
-		};
-
-		toggleLanguage();
-	}, []);
-
-	useEffect(() => {
 		const handleScroll = () => {
 			if (isHomePage) {
-				setIsScrolled(window.scrollY > window.innerHeight * 0.65);
+				setIsScrolled(window.scrollY > window.innerHeight * 0.7);
 			} else {
 				setIsScrolled(window.scrollY > 20);
 			}
@@ -220,8 +204,9 @@ const Navbar = () => {
 		<>
 			{/* Transparent Overlay (Only for Home Page) */}
 			{isHomePage && (
-				<div className="absolute left-0 top-5 w-full h-[18vh] bg-gradient-to-b from-primary via-[#7BA4D4] to-transparent z-1"></div>
+				<div className="absolute left-0 top-5 w-full h-[18vh] bg-gradient-to-b from-[#002147] via-[#4D79A6] to-transparent z-1"></div>
 			)}
+
 
 			{/* Fixed Navbar */}
 			<div className="fixed top-0 left-0 w-full z-50 transition-transform duration-200">
@@ -245,7 +230,7 @@ const Navbar = () => {
 					<a href="/">
 						<div className="flex items-center space-x-4">
 							<img src={imgSrc} alt="IIITN Logo" className="h-16 w-16" />
-							
+
 						</div>
 					</a>
